separate extension namespace
