+++ src/core/Client/RestClientException.cs
@@ -22,16 +22,5 @@ namespace BlazorFocused.Client
         /// <param name="message">Exception message</param>
         /// <param name="exception">Inner Exception</param>
         public RestClientException(string message, Exception exception) : base(message, exception) { }
-
-        /// <summary>
-        /// Initializes a new instance of <see cref="RestClientException"/> with the Http Request
-        /// information that caused the exception.
-        /// </summary>
-        /// <param name="method">Http method of request that caused exception</param>
-        /// <param name="httpStatusCode">Status code of the response from failed request</param>
-        /// <param name="url">Url of request that caused exception</param>
-        public RestClientException(HttpMethod method, HttpStatusCode httpStatusCode, string url) :
-            base($"{method} request failed with {httpStatusCode} at {url}")
-        { }
     }
 }
